Долг линта, порция 7: горячая перезагрузка в общих компонентах (#858)
Последнее правило долга — `react-refresh/only-export-components`, 40 вхождений.
Эта порция закрывает десять: всё, что в `shared/ui`. Модуль, экспортирующий и
компонент, и функцию, не может быть границей горячей подмены: правка любого его
байта не подменяется на месте, а поднимается вверх по импортам — до ближайшего
годного модуля.

Разделены семь файлов. Правило выбора, куда что уезжает: остаётся на прежнем
месте то, у чего больше потребителей, — иначе порция превращается в правку
импортов по всему интерфейсу. Поэтому у тостов и заголовка вкладки уехал
ПРОВАЙДЕР (его зовут из App.tsx один раз), а хуки остались, и двадцать два
импорта `useToast` не тронуты. У темы наоборот: файл назван по провайдеру, и
уводить провайдера из `ThemeProvider.tsx` значило бы лгать именем — уехал хук
(семь мест).

Три файла переименованы .tsx → .ts: расширения в путях импорта нет, поэтому
переименование ничего не стоит, а .tsx без JSX сбивает с толку.

Проверено, что правило вообще не пустое — разовой пробой (в харнесс не кладу,
она правит исходники): та же правка тела компонента при старом файле даёт
«invalidate … "useTheme" export is incompatible» и подъём до App.tsx, при новом
— подмену ровно этого модуля. Оговорка честности: состояние React переживало
правку в обоих случаях, потому что реестр Fast Refresh держит опознание
компонента; наблюдаемая плата — объём перевыполняемого, а не потерянный ввод.

Попутно: `.vscode/settings.json` уехал в master моим `git add -A` в прошлом PR —
снят с учёта, каталог внесён в .gitignore.

Проверено: tsc -b, сборка, 583 юнит-теста, храповик 40 → 30 и все семь живых
прогонов (fields 15, dialogs 13, settings 8, pages 8, shared-ui 6, keyboard 11,
routing 11).
