--- .vscode/settings.json
@@ -1,3 +1,0 @@
-{
-    "dotnet.preferCSharpExtension": true
-}
