+++ src/client/src/App.tsx
@@ -2,9 +2,9 @@ import { BrowserRouter, Routes, Route, Navigate } from 'react-router';
 import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
 import { ThemeProvider } from '@/shared/ui/ThemeProvider';
 import { ErrorBoundary } from '@/shared/ui/ErrorBoundary';
-import { NavigationGuardProvider } from '@/shared/ui/NavigationGuard';
-import { DocumentTitleManager } from '@/shared/ui/DocumentTitle';
-import { ToastProvider } from '@/shared/ui/Toast';
+import { NavigationGuardProvider } from '@/shared/ui/NavigationGuardProvider';
+import { DocumentTitleManager } from '@/shared/ui/DocumentTitleManager';
+import { ToastProvider } from '@/shared/ui/ToastProvider';
 import { BugReportProvider } from '@/shared/ui/BugReportProvider';
 import { AuthProvider } from '@/shared/ui/AuthProvider';
 import { ProtectedRoute, AdminRoute } from '@/shared/ui/ProtectedRoute';
