+++ src/client/src/features/account/ProfilePage.tsx
@@ -3,7 +3,8 @@ import { useServerForm } from '@/shared/hooks/useServerForm';
 import { useAccount, useUpdateAccount, useResendConfirmation, useChangeEmail, useUpdateAvatar } from '@/shared/api/account';
 import { TextField } from '@/shared/ui/TextField';
 import { Button } from '@/shared/ui/Button';
-import { Avatar, downscaleToDataUri } from '@/shared/ui/Avatar';
+import { Avatar } from '@/shared/ui/Avatar';
+import { downscaleToDataUri } from '@/shared/ui/avatarHelpers';
 import { ChangePasswordModal } from '@/shared/ui/ChangePasswordModal';
 import { apiError } from '@/shared/utils/apiError';
 import { KeyRound, CheckCircle2, AlertCircle, Mail, Upload, Trash2 } from 'lucide-react';
