+++ src/client/src/shared/ui/ListDetailShell.tsx
@@ -1,4 +1,4 @@
-import { useState, type ReactNode } from 'react';
+import type { ReactNode } from 'react';
 import { ChevronRight } from 'lucide-react';
 import { Button } from './Button';
 import { SearchInput } from './SearchInput';
@@ -157,30 +157,4 @@ export function DetailHeader({ heading, dirty, saving, onSaveAll, onRevert, acti
       </div>
     </div>
   );
-}
-
-/**
- * Гард несохранённых изменений при смене выбранного элемента. Generic по ключу выбора (`string` у типов
- * документов, `{mode,id}` у типов полей). Возвращает `request(next)` для перехвата выбора и `dialogProps`
- * для `LeaveGuardDialog`. `onCommit` применяет переход (страница владеет своим selectedKey).
- */
-export function useDirtyGuard<TKey>({ isDirty, saving, saveAll, onCommit }: {
-  isDirty: boolean; saving: boolean; saveAll: () => Promise<void>; onCommit: (next: TKey) => void;
-}) {
-  const [pending, setPending] = useState<{ next: TKey } | null>(null);
-  const request = (next: TKey) => { if (isDirty) setPending({ next }); else onCommit(next); };
-  const dialogProps = {
-    open: pending !== null,
-    saving,
-    onCancel: () => setPending(null),
-    onDiscard: () => { if (pending) onCommit(pending.next); setPending(null); },
-    onSave: async () => {
-      // При ошибке валидации/сохранения закрываем диалог, чтобы ошибка формы стала видна
-      // (иначе «Сохранить и перейти» молча висит и выглядит сломанным). Переход отменяется.
-      try { await saveAll(); if (pending) onCommit(pending.next); }
-      catch { /* ошибка уже показана в форме (setError) */ }
-      finally { setPending(null); }
-    },
-  };
-  return { request, dialogProps };
 }
